ci: modify CI/CD steps for EE
We avoid running jobs that will publish our product to public
repositories. This currently means that we cannot test cloud deployments
in an automated manner. However, none of our EE features are related to
cloud deployments so this should not be a problem.

(much later, during a rebase): also avoid publishing the algolia search
index from the EE docs, since we currently use the OSS index from both
docs sites.