fix: Initial prep_container should fail quickly (FOUNDENG-217) (#435)
Fix the exception rasied when master is not reacable (MasterNotFoundException)
APIHttpError only happens when call completes without a successful status
response.

The initial prep_container was ntended to fail reasonably quickly
to enable diagnostics of misconfiguration.   Recent changes
for re-using the common session (DET-8003) settings has caused the initial
communication to retry for more than 30 mins thus defeating the
original intent of prep_container.trial_prep and the error message
it provides.

This change lowers the session retry in prep_container (6 retries with 0.5
backoff -> 64 seconds) to enable the diagnostic message to be posted reasonably quickly.