fix: Starting state now shows Running [FOUNDENG-242] (#447)
Once the job was queue with Slurm/PBS we triggered the Starting
state.  Prior to 19.4 this use to show as QUEUED in the UI, but
now has changed to "RUNNING (PREPARING ENV)" which is not
accurate.   So map PENDING -> Assigned such that the UI
continues to show "QUEUED" until the job starts running.