chore: Add additional configuration options in slurmcluster.sh (#537)
Add the capability to set a default image, ask_container_defaults, and
partition_overrides in the master.yaml.

Add configuration for sawmil to make grizzly nodes cuda
and provide a default image, and MPI settings.

Eliminate the need for the CLUSTERS list to be manually updated by
just checking for the cluster configuration directly.

Indicate the generated master.yaml file name to simplify debugging
when injecting multiline options.