fix: Enable slots_per_node with gpu_type [FOUNDENG-401] (#623)
The tres_enabled:true case of slots_per_node is handled
in the dispatcher_task (not launcher).   When a gpu_type is
specified in a --gpus expression, it must also be applied
to the --gpus-per-task expression.

TODO:  Better if this were handled on the launcher side [FOUNDENG-403]

Tested on the grenoble system which surfaced the problem.