fix: Quick fix to reduce potential race conditions [FOUNDENG-447] (#698)
The launcher infrastructure uses file/directory management to track dispatch state and it has some design deficiencies that we has tickets to address.     We have previously been using a fixed name ("det") and
unspecified version ("unknown") which causes all dispatches from a user to end up in the same directories.    When there are lots of jobs coming and going we have observed file conflicts that causes data loss.

By providing a generated UUID as the version string, all dispatches for a user will be in a separate directories and therefore it should drastically reduce the potential for race conditions on file management until we can fix the launcher internals.