fix: Active job across upgrade 0.21.0-ee->0.22.2-ee causes crash [FE-156] (#998)
Prior to 0.22.2 the dispatchID was distinct from the allocationID and we
mapped between them.   In 0.22.2, we started using the allocationID
as the dispatchID.  What we missed is that active dispatches prior to
0.22.2 will not have this property so if we try to use the dispatchID
as the allocationID it will not find it.  After the upgrade to 0.22.2+
we lookup the AllocateRequest using the incorrect ID and it causes a
NPE.

This fix does the lookup of the AllocateRequest, and if it fails it
reverts to the old behavior of looking up the AllocationID from the
DB table.   This fix can be dropped when we stop supporting upgrades
from releases prior to 0.22.2