fix: Change Determined master to abort on bad launcher version [FE-184] (#1031)
To be sure that Determined will not be used with an incompatible
launcher version, immediately abort the server during startup
if the launcher version is insufficient or we cannot communicate with
the launcher.    The determined-master.service currently only retries the
startup for 20secs, this will be increase to avoid potential false failures
if systemd starts determined-master.service before launcher.service.